sprint-2/step-2 / fixed remarks
