+++ src/components/burger-constructor-item/index.tsx
@@ -8,8 +8,8 @@ import {
   ConstructorElement,
 } from "@ya.praktikum/react-developer-burger-ui-components";
 
-import { orderListSelector } from "store/orderList/selectors";
-import { replaceIngredient } from "store/orderList/slice";
+import { orderSelector } from "store/order/selectors";
+import { replaceIngredient } from "store/order/slice";
 
 import { IBurgerIngredientsItem } from "types/interfaces";
 
@@ -27,7 +27,7 @@ const BurgerConstructorItem = (props: IBurgerConstructorItemProps) => {
 
   const dispatch = useDispatch();
 
-  const { orderList } = useSelector(orderListSelector);
+  const { orderList } = useSelector(orderSelector);
 
   const ref = useRef<HTMLElement | null>(null);
 
@@ -39,11 +39,26 @@ const BurgerConstructorItem = (props: IBurgerConstructorItemProps) => {
         handlerId: monitor.getHandlerId(),
       };
     },
-    hover(ingredient: any) {
-      if (!ref.current) return;
-
+    hover(ingredient: any, monitor) {
       const dragIndex = ingredient.index;
       const hoverIndex = index;
+
+      if (dragIndex === hoverIndex) return;
+
+      const hoverBoundingRect = ref.current?.getBoundingClientRect();
+
+      if (!hoverBoundingRect) return;
+
+      const hoverMiddleY =
+        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
+      const clientOffset = monitor.getClientOffset();
+
+      if (!clientOffset) return;
+
+      const hoverClientY = clientOffset.y - hoverBoundingRect.top;
+
+      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) return;
+      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) return;
 
       if (
         ingredient.item.type !== "bun" &&
@@ -51,6 +66,8 @@ const BurgerConstructorItem = (props: IBurgerConstructorItemProps) => {
       ) {
         dispatch(replaceIngredient({ dragIndex, hoverIndex }));
       }
+
+      ingredient.index = hoverIndex;
     },
   });
