+++ src/components/header/index.tsx
@@ -11,19 +11,23 @@ const AppHeader = () => {
   return (
     <header className={styles.header}>
       <nav>
-        <BurgerIcon type="primary" />
-        <span
-          className={`${styles.headerItemText} text text_type_main-default mr-10 ml-2`}
-        >
-          Конструктор
-        </span>
+        <a href="#" className={styles.navItem}>
+          <BurgerIcon type="primary" />
+          <span
+            className={`${styles.headerItemText} text text_type_main-default mr-10 ml-2`}
+          >
+            Конструктор
+          </span>
+        </a>
 
-        <ListIcon type="secondary" />
-        <span
-          className={`${styles.headerItemText} text text_type_main-default ml-2 text_color_inactive`}
-        >
-          Лента заказов
-        </span>
+        <a href="#" className={styles.navItem}>
+          <ListIcon type="secondary" />
+          <span
+            className={`${styles.headerItemText} text text_type_main-default ml-2 text_color_inactive`}
+          >
+            Лента заказов
+          </span>
+        </a>
       </nav>
 
       <nav>
@@ -31,12 +35,14 @@ const AppHeader = () => {
       </nav>
 
       <nav>
-        <ProfileIcon type="secondary" />
-        <span
-          className={`${styles.headerItemText} text text_type_main-default ml-2 text_color_inactive`}
-        >
-          Личный кабинет
-        </span>
+        <a href="#" className={styles.navItem}>
+          <ProfileIcon type="secondary" />
+          <span
+            className={`${styles.headerItemText} text text_type_main-default ml-2 text_color_inactive`}
+          >
+            Личный кабинет
+          </span>
+        </a>
       </nav>
     </header>
   );
