--- src/store/modal/selectors.ts
@@ -1,3 +1,0 @@
-import { RootState } from "store";
-
-export const modalSelector = (state: RootState) => state.modal;
