+++ src/store/order/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from "store";
+
+export const orderSelector = (state: RootState) => state.order;
