--- src/store/orderList/selectors.ts
@@ -1,3 +1,0 @@
-import { RootState } from "store";
-
-export const orderListSelector = (state: RootState) => state.orderList;
