+++ src/utils/burger-API.ts
@@ -1,4 +1,4 @@
-export const apiURL = "https://norma.nomoreparties.space";
+export const API_URL = "https://norma.nomoreparties.space";
 
 export const checkResponse = (response: Response) =>
   response.ok
